modules/module-05-data-lakehouse/exercises/01-delta-basics/starter/requirements.txt
# Requirements for Exercise 01: Delta Basics

# Core Spark and Delta Lake
pyspark==3.5.0
delta-spark==3.0.0

# AWS S3 support (for MinIO)
boto3==1.34.34
s3fs==2024.2.0

# Testing (if running local tests)
pytest==8.0.0
pytest-spark==0.6.0
